feat: update handling money, change float64 to int64

+++ Repository/account_repository.go
@@ -47,7 +47,7 @@ func (r *AccountRepository) GetAccountForUpdate(
 func (r *AccountRepository) DeductBalance(
 	tx *sql.Tx,
 	accountID string,
-	amount float64,
+	amount int64,
 ) error {
 
 	_, err := tx.Exec(
@@ -64,7 +64,7 @@ func (r *AccountRepository) DeductBalance(
 func (r *AccountRepository) AddBalance(
 	tx *sql.Tx,
 	accountID string,
-	amount float64,
+	amount int64,
 ) error {
 
 	_, err := tx.Exec(
