+++ model/account.go
@@ -3,17 +3,17 @@ package model
 import "time"
 
 type Account struct {
-	ID        string  `json:"id"`
-	OwnerName string  `json:"owner_name"`
-	Balance   float64 `json:"balance"`
-	Status    string  `json:"status"`
+	ID        string `json:"id"`
+	OwnerName string `json:"owner_name"`
+	Balance   int64  `json:"balance"`
+	Status    string `json:"status"`
 }
 
 type TransferRequest struct {
-	ID            string  `json:"id"`
-	FromAccountID string  `json:"from_account_id"`
-	ToAccountID   string  `json:"to_account_id"`
-	Amount        float64 `json:"amount"`
+	ID            string `json:"id"`
+	FromAccountID string `json:"from_account_id"`
+	ToAccountID   string `json:"to_account_id"`
+	Amount        int64  `json:"amount"`
 }
 
 type Transfer struct {
@@ -21,8 +21,7 @@ type Transfer struct {
 	ReferenceNumber string    `json:"reference_number"`
 	FromAccountID   string    `json:"from_account_id"`
 	ToAccountID     string    `json:"to_account_id"`
-	Amount          float64   `json:"amount"`
+	Amount          int64     `json:"amount"`
 	Status          string    `json:"status"`
 	CreatedAt       time.Time `json:"created_at"`
 }
-
